Code/Cylinder3D-desnowing/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.3.29=h06a4308_0
ccimport=0.3.7=pypi_0
certifi=2021.10.8=py37h06a4308_2
charset-normalizer=2.0.12=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cumm-cu102=0.2.8=pypi_0
cython=0.29.28=pypi_0
ffmpeg=4.3=hf484d3e_0
fire=0.4.0=pypi_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
idna=3.3=pypi_0
importlib-metadata=4.11.3=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9d=h7f8727e_0
lame=3.100=h7b6447c_0
lark=1.1.2=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libllvm11=11.1.0=h3826bc1_1
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
llvmlite=0.38.0=py37h4ff587b_0
lz4-c=1.9.3=h295c915_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
ninja=1.10.2.3=pypi_0
numba=0.55.1=py37h51133e4_0
numpy=1.21.5=py37he7a7128_1
numpy-base=1.21.5=py37hf524024_1
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h7f8727e_0
pccm=0.3.4=pypi_0
pillow=9.0.1=py37h22f2fdc_0
pip=21.2.2=py37h06a4308_0
portalocker=2.4.0=pypi_0
pybind11=2.9.2=pypi_0
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytorch=1.10.1=py3.7_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytorch-scatter=2.0.9=py37_torch_1.10.0_cu102
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pypi_0
setuptools=61.2.0=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
spconv-cu102=2.1.21=pypi_0
sqlite=3.38.2=hc218d9a_0
strictyaml=1.6.1=pypi_0
tbb=2021.5.0=hd09550d_0
termcolor=1.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.10.1=py37_cu102
torchvision=0.11.2=py37_cu102
tqdm=4.64.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zipp=3.8.0=pypi_0
zlib=1.2.12=h7f8727e_1
zstd=1.4.9=haebb681_0
